Refactor bitwise ops on booleans and remove duplicated `Set` from BinOp (#4728)
This PR makes a few changes to how bitwise and boolean ops are computed
in the arkouda server.

- Since the order of operands to bitwise ops does not matter, by using a
helper function `doBoolBoolBitOp` and only calling it with operands in
the same order (always `vector, scalar` or `vector, vector`) we can save
some compilation time
- The same `Set.set` was being recreated all over the place to determine
if an op was an operator. This PR refactors that set into a common place
so it does not get instantiated for all the generic combinations.
- Makes a similar transformation with `Set.set` for `realOps`

I consider this a general cleanup and consolidation of logic in
BinOp/OperatorMsg, it also saves a little bit of time while compiling

---------

Signed-off-by: Jade Abraham <jade.abraham@hpe.com>