Closes #4606 benchmark v2/reformat benchmark results.py to parse all benchmark results (#4751)
This PR moves the generation of field lookup map to a separate script
`benchmark_v2/generate_field_lookup_map.py`. This will enable the map to
be regenerated only when there are major code changes, rather than with
each benchmark run.

It also adds field parsing for most of the benchmarks. The few remaining
(flatten, split, multiIO) will be handled in a separate PR as they
require extra work.

Additionally, the `tranfer_rate` field was changed to float (rather than
string) type in order to match the form expected in the results.

Closes #4606 benchmark v2/reformat benchmark results.py to parse all
benchmark results

Co-authored-by: ajpotts <ajpotts@users.noreply.github.com>