Add heartbeat on the server (#4750)
This heartbeat allows client to recognize that the connection to the
server went away while the server was processing a message from the
client. The client raises an exception in this case. The heartbeat is
activated by passing a positive number to the `timeout` argument of
`ak.connect()`.

While there, tidy up some code. Also, introduce
`ak.client.server_sleep(seconds)` to enable testing of the heartbeat,
and `ak.client.note_for_server_log(message)` to make it easier to track
the tests being tested while examining `arkouda.log`.

Implementation-wise, this enables "TCP keepalive" on the ZMQ socket
between the client and the server and adds a monitor for this socket on
the client. After sending a request to the server, the client waits for
a reply with a timeout. If it times out, the client checks the monitor
for "closed "or "disconnected" events. If none, the client goes back to
waiting for a server reply with a timeout. As before, there is no
overall timeout for server response.

While there, moved `test_no_op` and `test_ruok` within `client_test.py`
to match the relative position of `_no_op` and `ruok` within
`client.py`.

Related: #1010

---------

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>