Closes #4498: Create a repartitionByLocale function (#4647)
Created the repartitionByLocale function ~~changing the behavior of
flatten to demonstrate as a proof of concept. Seems to pass multilocale,
multidimensional tests~~.

Previously I was using flatten to demonstrate it worked, and it did work
and pass tests there. Since it would be difficult to directly test this
code otherwise, I think this should be treated as passing tests.

Closes #4498: Create a repartitionByLocale function