Add SegStringSymEntry's offsets and values to SymTab
Rationale: I realized that having them in SymTab will significantly
simplify the implementation of checkpointing for SegStringSymEntry.

I came across Registry.chpl and observed that ObjType.STRINGS things
are handled like ObjType.PDARRAY. On the surface STRINGS need to be
treated like ObjType.SEGARRAY instead because they have two components,
one of which is probably forgotten about when handling a STRINGS
as a PDARRAY. Thankfully this did not affect me.

---
Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>