Checkpoint pdarrays of bigints (#4771)
Add support for bigint arrays to checkpointing.

The initial implementation used ULEB128 encoding for saving "count",
i.e., the number of bytes needed to store a given bigint. The intention
was to use the smallest number of bytes possible for representing the
"count", as ULEB128 is variable-length, akin to utf-8. I changed to
always using 4 bytes for saving "count" in the second commit to simplify
the code. I consider the use of extra 2 bytes for bigints that fit
within (2^15-1) bytes (4 bytes vs. ULEB128) to be a reasonable
compromise between maintainability and efficiency.

Contributes to #2384.

---------

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>