4688 intermittent ci failures error invoking driver compilation phase (#4776)
The purpose of this PR is to reduce the RAM usage for the
multi-dimensional unit tests running the CI. It adds a script that will
comment out a list of modules from the `ServerModules.cfg` that do not
need to be tested in the multi-dimensional case. It also adds a pytest
skip to any unit tests that depend on these modules when arkouda is
compiled for the multidimesional case.

Co-authored-by: ajpotts <ajpotts@users.noreply.github.com>