ak.get_registration_config() and tweaks (#4752)
* Add `ak.get_registration_config()`.
* Populate a Generator in the client with GeneratorSymEntry instances on
the server only for those types that are requested in
`registration-config.json`.
* Minor code improvements I did while working on the following.

While there, redirected the file naming for bigint checkpoint data to
the standard-ish `generateFilename()`.

This PR initially added checkpointing of `GeneratorSymEntry`. However
during testing I realized that, while checkpointing per se works,
restoring from a checkpoint does not "roll back" Arkouda properly. This
is because part of the Generator state, specifically its `state` field,
is stored in the client and not on the server. Therefore backed out this
checkpointing addition in the second commit in this PR.

See #4709 for discussion on this.

---------

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>