Closes 4831, standardizes axis checking for cases where axis can only be an integer (#4834)
Closes #4831, which is a subset of #4800

This covers the cases where the axis parameter can only be an integer.
This only applies to array_api and numpy functions, and not the pandas
functions (where axis can be 0 or 1, "index" or "columns").

It doesn't cover the cases where the axis parameter can be a tuple.

Quick summary:

There is now a function in numpy/util.py called
_integer_axis_validation. It returns a bool and an int. If the bool is
false, the given axis wasn't valid, and in (I think) every case, an
IndexError is raised.

There are several functions in array_api that just pass their parameters
to arkouda/numpy functions with the same name. In those cases, I don't
bother to check at the array_api level.

Note that array_api's vecdot accepted an axis parameter, but didn't use
it. I changed that so that it passes the axis argument to the numpy
function. Also note that vecdot handles the axis validation differently
than other functions, so it doesn't use the standard validation.

Also also note that expand_dims and stack are unique in that they check
the validity against a rank 1 greater than the given input.

Co-authored-by: drculhane <drculhane@users.noreply.github.com>