Incorporates standard for axis validation to array_api/manipulation_functions.py (#4858)
Addendum

The existing axis checking in array_api **moveaxis** wasn't correct. It
allowed cases numpy doesn't (for example, moving axes (0,1,2) to
(1,1,0)). This isn't the same issue as the axis validation, but it was
an error, and has been fixed.

Closes #4857 

This is part of a series of PRs that will standardize the axis checking
(and negative axis handling) in arkouda. This PR specifically finalizes
the _axis_validation function in numpy/util.py, and updates all of the
axis handling and checking in array_api/manipulation_functions.py.

This PR also adds several fields to the docstrings, and changes the
"must have same rank" check in **stack** to "must have same shape," to
match numpy.

Co-authored-by: drculhane <drculhane@users.noreply.github.com>