Fixes handling of size/shape parameter in ak.random.poisson (#4916)
Closes #4220

This fixes the handling of the size parameter in ak.random.poisson, so
that (e.g.) ak.random.poisson(lam=lam,size=(m,n)) will produce an mxn
pdarray.

The very simple fix is to generate 1D values, and reshape if needed.

Analyzing this bug report uncovered other issues in the random number
generation, so I'll be adding additional issues.

---------

Co-authored-by: drculhane <drculhane@users.noreply.github.com>