Standardizes axis validation and handling in array_api/statistical_functions.py (#4909)
Closes #4905 

The primary intent of this PR was to add the new standard axis
validation & handling to array_api/statistical_functions.py. Along the
way, I noted other issues.

- The docstrings were very incomplete.  I have attempted to fix that.
- The unit test for array_api's mean had a print statement where there
should have been an assert. Fixed.
- There are two comments in statistical_functions.py that I believe are
out of date: lines 298 and 449, if my notes are correct. I'm inclined to
delete them (holding off on that for a later revision).

NOTE: some of the functions are just pass-throughs (i.e. no code except
for calling another function). In those cases, I've chosen not to check
the axis parameter (since the other function has to), nor to include the
corresponding errors in the Raises section, even though I may have
included Examples. This is purely a judgment call, and I'm open to
discussion.

Co-authored-by: drculhane <drculhane@users.noreply.github.com>