Improve parquet read implementation and multiple column read performance (#4906)
This PR has two main goals:

- improve the performance of parquet reads, particularly when all the
columns of the file(s) are read
- start improving the implementation on C++ and Chapel sides

### User-facing Change:

- This PR deprecates `has_non_float_nulls: Bool`, in favor of
`null_handling: {"none", "only floats", "all"}`
- We need a 3-way switch to enable fast floating point reads when the
user knows their data contains no nulls. Old boolean-based switch cannot
capture that. The new `null_handling` formal to `read_parquet` defaults
to "only floats" to match with the older behavior.
- Right now, `null_handling` is effective only when reading all columns.
`has_non_float_nulls` is used when only a subset of columns are read. It
is a TODO to phase out `has_non_float_nulls`, among other things.

### Why do we need a new implementation for performance?

Parquet is a columnar format. However, that doesn't mean that all data
for a column is stored consecutively. For large files (512M by default,
I think), Parquet files will be written in consecutive _row groups_
instead of storing large columns one after another. Therefore, a more
natural way of reading a parquet file is to read row-group-by-row-group
while going through all the columns. Considering reading all columns as
a common use case, it is important to read parquet files in that way for
performance. Right now, if you have 10 columns, you open and read the
entire file(s) 10 times, one for each column, skipping 9 columns during
each read.

#### A small optimization along the way

With the new `null_handling` control flag, we can also enable batch
reads of floating point types where prior to this PR we were reading
them one-by-one to be able to catch nulls. I think even with nulls at
play, we should be able to do better than that, but that's a post-merge
TODO.



#### Performance Improvements

This is using ~400GBs of data in 5 columns partitioned into 128 files,
read by 16 locales:

|           | `main` (s) | branch (s) | improvement (x)
|---------|-----------------|---------------|--------
| min    | 11.43           | 7.60          | 1.50
| max   | 22.58          | 14.96        | 1.51
| ave    |  16.69         | 9.38          | 1.78

This is using ~30GBs of data in 1000 columns partitioned into 128 files,
read by 16 locales:

|           | `master` (s) | branch (s) | improvement (x)
|---------|-----------------|---------------|--------
| min    | 335.97           | 9.02          | 37.24
| max   | 335.97          | 15.64        | 21.48
| ave    | 335.97         | 9.98          | 33.67



### Implementation details:

- If the user doesn't pass any dataset names to `read_parquet`, we read
all columns. This PR makes that case take an entirely new codepath with
no modifications to the subset-of-columns code path. I intend to remove
this old code path in favor of the new one even for subset-of columns.
That's a post-merge TODO.
- This new code path introduces the following on the C++ side:
- `namespace akcpp` on the C++ side. I want more code to live in this
namespace going forward. Within that namespace:
- `struct AkReadColOp` on the C++ side to represent type-agnostic
information for reading a single column.
- This type has a template `read()` function for reading a column with
specific dynamic type
- `read()` has some template specializations for handling some specific
cases like short integrals that are not represented on the Chapel side
and need to be cast to bigger ones
- The `read()` family of functions has _a lot of code redundancy_ that I
inherited from the earlier implementation. Refactoring those
redundancies into helpers is an obvious TODO.
- `readAllCols` function to handle the operation. This is the function
that selects the correct `read()` instantiation based on the dynamic
type.
- This new code path introduces the following on the Chapel side:
- `readAllColsParquetMsg` which is the main message handler here. The
implementation is too much of a Frakenstein experiment where old
style/structure is mixed with the new. Going forward, I want this
function to be the general parquet read function, so it needs more
improvements.
- The implementation of this stores read metadata in `pqReadColOp`. This
is deceptively similar to `AkReadColOp`. But it actually stores
information for the whole read operation and it is not per-column. I
want to consider whether we can have single metadata shared between C++
and Chapel.
- This `pqReadColOp` determines whether we can read all columns at the
same time based on the dynamic types of the columns. If we can't it will
fall back to the old implementation. Currently that fallback is a bit
rough as we call the message handler. I think that's a bad practice and
the core logic of the read has to be available without the message
handling. We'll need to reconcile the two different reads in any case,
so that's part of post-merge TODOs.
- I used this type to do column-per-column reads, too. That's the only
change in the existing code path. I started off thinking that I can do
all the refactoring in one go, but decided against it. Nevertheless, I
left that part in. Currently that causes `pqReadColOp` to be more
generic than I'd like. It will be fixed in the next refactor wave.
  - Simplified error location reporting
  - Several new small helpers and simplified helpers 

### Test Status:
- [x] local parquet unit testing
- [x] oversubscribed parquet unit testing

### Post-Merge TODO
- [ ] Refactor `read()` functions into having more helpers
- [ ] Consider handling per-batch post-processing differently, maybe
with functors.
- [ ] use the `akcpp::readAllCols` function for reading subset of
columns, too and remove the old read implementation completely.
Similarly for `readAllColsParquetMsg` on Chapel side
  - [ ] Fall back to col-per-col read in a more graceful way 
  - [ ] completely deprecate `has_non_float_nulls`
- [ ] Can we read floating point columns with nulls in them in batches
and process nulls differently?
- [ ] Can we use a single metadata type for C++ and Chapel?
- [ ] Use message registration attributes for Parquet message handlers
- [ ] Short integer and float reads use a temporary buffer. We can get
rid of that by reading into the larger Chapel buffer and spacing things
out in reverse order.
- [ ] Handle `ARROWERROR` and error messages populated by the C level
better and more consistently
- [ ] https://github.com/Bears-R-Us/arkouda/issues/4925

Signed-off-by: Engin Kayraklioglu <e-kayrakli@users.noreply.github.com>