Part 3 of: #4348 Remove DAR103 errors (#4992)
This PR moves to per-file ignores for docstring related error codes. The
purpose it to make it easier to resolve the errors one module at a time
and break the resolution process in to manageable chunks.

Also removes `darglint` CI test, because it is redundant with `flake8`,
which already catches these errors.

Co-authored-by: ajpotts <ajpotts@users.noreply.github.com>