Closes #4227: Hitting Chapel's default instantiation limit when compiling for 3 or more dimensions (#5002)
This allows users to override Chapel's instantiation limit from the make
line.

As it happens, this is not easy to test that it "fixes" anything right
now. I think recent PRs have done an outstanding job of fixing this
issue from the perspective of not needing such a high instantiation
limit (and, I might add, the memory issue). On my laptop with about 30
GB RAM free, I was able to set

```yaml
# Snippet from registration_config.json
"nd": [1, 2, 3, 4, 5],
```

(yes, five dimensions) and while it took a very long time to compile
(around 40 minutes) it managed to compile with the default instantiation
limit of 512. So how do we know that this works? It's actually not that
hard to check that it breaks things. You can run

```bash
make instantiate_max=10
```

and you will see

```bash
$CHPL_HOME/modules/standard/CTypes.chpl:611: error: Function ':' has been instantiated too many times
note:   If this is intentional, try increasing the instantiation limit from 10
make: *** [Makefile:575: arkouda_server] Error 1
```

So clearly this code has an effect! That really ought to mean that if a
user wants a higher instantiation limit, they should be able to get it.

Oh, and I updated `BUILD.md` so anyone who greps through our files for
the error should be able to find documentation on how to fix it!

Closes #4227: Hitting Chapel's default instantiation limit when
compiling for 3 or more dimensions