Closes #4500: Create a repartitionByHash function (#5005)
Since deciding to switch over to the `innerArray` framework, I figured I
would do this as an `innerArray` version. I also did the same for the
`Strings` version.

I'm thinking I might change the way these work at some point. It'd be
nice if I could do the hashing inside the first coforall of the function
it moves on to. I've been thinking about how to make that work and I
think it would maybe be a little tricky. But I think I might be able to
do it with defining classes or records or something. And then that would
allow users to define their own repartitioning schemes as well.

Closes #4500: Create a repartitionByHash function