Add numNodes to server config to support colocale-based runs (#5021)
This PR adds a `numNodes` key to the server config. With Chapel's
colocale support, number of locales does not always equal to number of
nodes. Especially when running performance benchmarks, it is important
to be able to fix the data size based on number of nodes, and not
locales, as the former is what represents amount of physical resources
we have.

I will create a follow up PR to adjust performance benchmarking to use
number of nodes instead of locales shortly.

---------

Signed-off-by: Engin Kayraklioglu <e-kayrakli@users.noreply.github.com>