Temporarily disable pytables installation test in the CI (#5036)
Temporarily disable the test of the `pytables` install in the
`Makefile`. This will give time to investigate the error. It's also
possible that `pip` install is now available for the versions we need,
in which case we may not need to install `pytables` via the `Makefile`
going forward.

Related to #5033 intermittent arkouda makefile test failures

Co-authored-by: ajpotts <ajpotts@users.noreply.github.com>