Closes 5040, fixing a bug in multi-dim matrix multiplication (#5042)
Closes #5040.

As written, it was requiring matrices a and b to have the same domain,
which was incorrect. The actual requirements are: (1) they have the same
rank, and (2) the last two dimensions of each must be compatible with 2D
matrix multiplication, i.e. m,k and k,n.

This fixes that bug.

Note about changes:

chapel-side: b now gets its own domain (named db), and the where clause
in the proc includes `da.rank == db.rank`

python-side: the generic_msg call to the chapel-side now includes the
rank of the right-side matrix (because the interface requires it), and
`pdaLeft` and `pdaRight` have been renamed to `pda_L` and `pda_R`,
largely because some lines were going over the 105 character limit.

---------

Co-authored-by: drculhane <drculhane@users.noreply.github.com>