Closes #5045:  test_is_sorted fails when size<=10 (#5046)
In `test_is_sorted`, and `test_is_locally_sorted`, a random vector of
size `size // 10` and is expected to not be sorted. However, when
`size<=10`, the vector is degenerate and will automatically be sorted,
hence the test automatically fails. This PR implements a simple fix of
skipping that logic when `size < 100`.

Closes #5045:  test_is_sorted fails when size<=10

Co-authored-by: ajpotts <ajpotts@users.noreply.github.com>