Closes #5117: Unary Negation on ak.uint64 Does Not Match Numpy Semantics (#5124)
Since both of these are unary operators and in light of the testing
framework in PR #5121, I thought I would bundle both the negation
changes for `uint64` and `bool` together.

For unary negation on `uint64`, twos complement suggests that the way to
handle this is to negate the bits and add 1.

For unary negation on `bool`, the solution suggested was to raise a
`TypeError` and suggest using `~` instead or `logical_not`, but since
`logical_not` did not exist, I added it and have it pointing at `~`
since they seem to largely be the same thing. I think additional tests
specific to `logical_not` would be unnecessary if `~` is being tested
because it's literally a wrapper for `~`. However, I did grab the unit
tests from PR #5121 that seem relevant to this PR.

Closes #5117: Unary Negation on ak.uint64 Does Not Match Numpy Semantics
Closes #5119: Bug: Unary Negation (-a) on ak.bool_ Does Not Match NumPy
Semantics

Co-authored-by: ajpotts <amanda.j.potts@gmail.com>