Closes 5172 -- adds overloads to hash in numeric.py (#5217)
Closes #5172.

This adds overloads to the `hash` function in _numeric.py_, implements
typechecking in `hash`, removes the `type: ignore` comment, and removes
the comment describing that the function was not being typechecked.

It does not add overloads to the hash functions in _categorical,
segarray or strings_, because all of those are of the form `def hash
(self)`.

IMHO, the overloads are a bit wordy, but this is what mypy requires. I
did consult chatGPT, which did not recommend breaking out the elements
of the Union, since that aspect doesn't affect the output type. Only the
value of `full` does.

Co-authored-by: drculhane <drculhane@users.noreply.github.com>